requirements.txt
bs4
html5lib
groupy
python-dateutil
pyyaml
